requirements.txt
Django==1.11.5
django-extensions==1.9.1
django-rest-framework==0.1.0
djangorestframework==3.6.4
pkg-resources==0.0.0
pytz==2017.2
RPi.GPIO==0.6.3
six==1.11.0
